pyproject.toml
[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[project]
name = "create-issue"
version = "0.3.4"
description = "Create GitLab issues (and optionally merge requests) from the CLI"
dependencies = ["python-gitlab"]

[project.scripts]
create-issue = "create_issue.cli:main_cli"

